+++ packages/components-angular/src/app/pages/index.ts
@@ -28,6 +28,7 @@ import { BannerComponent } from './banner.component';
 import { ButtonComponent } from './button.component';
 import { ButtonGroupComponent } from './button-group.component';
 import { ButtonPureComponent } from './button-pure.component';
+import { ButtonTileComponent } from './button-tile.component';
 import { CarouselComponent } from './carousel.component';
 import { CheckboxWrapperComponent } from './checkbox-wrapper.component';
 import { ContentWrapperComponent } from './content-wrapper.component';
@@ -84,6 +85,7 @@ export const generatedPages = [
   ButtonComponent,
   ButtonGroupComponent,
   ButtonPureComponent,
+  ButtonTileComponent,
   CarouselComponent,
   CheckboxWrapperComponent,
   ContentWrapperComponent,
@@ -160,6 +162,11 @@ export const generatedRoutes: ExtendedRoute[] = [
     name: 'Button Pure',
     path: 'button-pure',
     component: ButtonPureComponent,
+  },
+  {
+    name: 'Button Tile',
+    path: 'button-tile',
+    component: ButtonTileComponent,
   },
   {
     name: 'Carousel',