+++ packages/storefront/tests/e2e/specs/dead-link-checker.e2e.ts
@@ -41,7 +41,7 @@ const internalUrls = getInternalUrls();
 const externalUrls = getExternalUrls();
 
 it('should have no exponential increase in internal urls', () => {
-  expect(internalUrls.length).toBeLessThanOrEqual(220);
+  expect(internalUrls.length).toBeLessThanOrEqual(250);
 });
 
 it.each(internalUrls.map<[string, number]>((url, i) => [url, i]))(