+++ packages/components-angular/tests/vrt/specs/styles/focus.vrt.ts
@@ -1,8 +1,15 @@
 import { extendedViewports, getVisualRegressionTester, vrtTest } from '@porsche-design-system/shared/testing';
+// TODO: we should move CDP to either shared package or maybe even better to visual-regression-tester itself
+import { forceFocusState } from '../../../../../components-js/tests/vrt/puppeteer/helpers';
 
 const id = 'styles-focus';
-// TODO: we need to force focus state by CDP
-// TODO: Angular/SCSS version renders slightly different compared to React/JSS
-xit.each(extendedViewports)('should have no visual regression for viewport %s', async (viewport) => {
-  expect(await vrtTest(getVisualRegressionTester(viewport), id, `/${id}`)).toBeFalsy();
+it.each(extendedViewports)('should have no visual regression for viewport %s', async (viewport) => {
+  expect(
+    await vrtTest(getVisualRegressionTester(viewport), id, `/${id}`, {
+      scenario: async (page) => {
+        await forceFocusState(page, 'a');
+        await forceFocusState(page, 'button');
+      },
+    })
+  ).toBeFalsy();
 });