+++ packages/components-vue/src/pages/OverviewPage.vue
@@ -5,6 +5,7 @@
     PButton,
     PButtonGroup,
     PButtonPure,
+    PButtonTile,
     PCarousel,
     PCheckboxWrapper,
     PContentWrapper,
@@ -313,6 +314,17 @@
         </PLinkTile>
       </div>
 
+      <div class="playground light" title="should render default button-tile">
+        <PButtonTile :label="'Some Label'" :description="'Default'">
+          <img
+            src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
+            width="50"
+            height="50"
+            alt="Beach"
+          />
+        </PButtonTile>
+      </div>
+
       <div class="playground light" title="should render default carousel">
         <PCarousel :heading="'Heading'">
           <div>Slide 1</div>
@@ -577,6 +589,17 @@
           </PLinkTile>
         </div>
 
+        <div class="playground light" title="should render default button-tile with custom prefix">
+          <PButtonTile :label="'Some Label'" :description="'Default'">
+            <img
+              src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
+              width="50"
+              height="50"
+              alt="Beach"
+            />
+          </PButtonTile>
+        </div>
+
         <div class="playground light" title="should render default carousel with custom prefix">
           <PCarousel :heading="'Heading'">
             <div>Slide 1</div>
@@ -618,7 +641,9 @@
   }
 
   p-link-tile,
-  my-prefix-p-link-tile {
+  p-button-tile,
+  my-prefix-p-link-tile,
+  my-prefix-p-button-tile {
     max-width: 400px;
   }