+++ packages/components/CHANGELOG.md
@@ -283,6 +283,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `Tabs Bar` and `Tabs` prop `weight` got value `semi-bold` and deprecated `semibold` value
   - `Tag` prop `color` got values `notification-info-soft`, `notification-warning-soft`, `notification-success-soft`,
     `notification-error-soft` and deprecated `notification-warning`, `notification-success`, `notification-error` values
+- `Button Tile` [PR](https://github.com/porsche-design-system/porsche-design-system/pull/2381).
 
 #### Changed